Extract internal API out of `ActiveWatch` into new `impl.EventHandlingWatch`
