+++ src/main/java/engineering/swat/watch/impl/jdk/JDKBaseWatch.java
@@ -37,10 +37,10 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.checkerframework.checker.nullness.qual.Nullable;
 
-import engineering.swat.watch.ActiveWatch;
 import engineering.swat.watch.WatchEvent;
+import engineering.swat.watch.impl.EventHandlingWatch;
 
-public abstract class JDKBaseWatch implements ActiveWatch {
+public abstract class JDKBaseWatch implements EventHandlingWatch {
     private final Logger logger = LogManager.getLogger();
 
     protected final Path path;
@@ -116,7 +116,7 @@ public abstract class JDKBaseWatch implements ActiveWatch {
         throw new IllegalArgumentException("Unexpected watch kind: " + jdkKind);
     }
 
-    // -- ActiveWatch --
+    // -- EventHandlingWatch --
 
     @Override
     public void handleEvent(WatchEvent e) {
